Django/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.8
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-environ==0.8.1
django-extensions==3.1.5
django-heroku==0.3.1
django-qr-code==2.3.0
django-redis==5.0.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
msgpack-python==0.5.6
Naked==0.1.31
psycopg2==2.9.2
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
pyOpenSSL==21.0.0
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
segno==1.4.1
service-identity==21.1.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.0
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.13.3
zope.interface==5.4.0
